fix: js-scope refactor - various renames and structure changes
