+++ src/main/scopes/jsx/interfaces.ts
@@ -6,8 +6,6 @@
  */
 
 import { ComplexValue } from '../js/complex-value.js'
-import { JsImport } from '../js/interfaces.js'
-import { PackageData } from '../npm/interfaces.js'
 
 export interface JsxElementAttribute {
   name: string
@@ -18,25 +16,4 @@ export interface JsxElement {
   name: string
   prefix: string | undefined
   attributes: JsxElementAttribute[]
-}
-
-export interface FileTree {
-  path: string
-  children: FileTree[]
-}
-
-export interface JsxElementImportMatcher {
-  findMatch: (element: JsxElement, imports: JsImport[]) => JsImport | undefined
-}
-
-export interface DependencyTreeDependency {
-  version: string
-  dependencies: Record<string, DependencyTreeDependency>
-}
-
-export interface DependencyTree extends PackageData {
-  name: string
-  version: string
-  dependencies: Record<string, DependencyTreeDependency>
-  [key: string]: unknown
 }
