+++ src/main/scopes/npm/get-package-sub-tree.ts
@@ -8,8 +8,7 @@
 import getPropertyByPath from 'lodash/get.js'
 
 import { InvalidObjectPathError } from '../../exceptions/invalid-object-path-error.js'
-import { DependencyTree } from '../jsx/interfaces.js'
-import { type InstallingPackage, type PackageData } from './interfaces.js'
+import { DependencyTree, type InstallingPackage, type PackageData } from './interfaces.js'
 
 /**
  * Given a dependency tree and an object path within that tree, return an object containing name,
