+++ src/test/core/get-tracked-source-files.test.ts
@@ -7,9 +7,9 @@
 import path from 'path'
 import { describe, expect, it } from 'vitest'
 
-import { getTrackedSourceFiles } from '../../../../main/scopes/jsx/utils/get-tracked-source-files.js'
-import { Fixture } from '../../../__utils/fixture.js'
-import { initLogger } from '../../../__utils/init-logger.js'
+import { getTrackedSourceFiles } from '../../main/core/get-tracked-source-files.js'
+import { Fixture } from '../__utils/fixture.js'
+import { initLogger } from '../__utils/init-logger.js'
 
 describe('getTrackedSourceFiles', () => {
   const logger = initLogger()
@@ -17,7 +17,13 @@ describe('getTrackedSourceFiles', () => {
   it('correctly returns all tracked source files', async () => {
     const root = new Fixture('projects/all-extensions')
 
-    const sourceFiles = await getTrackedSourceFiles(root.path, logger)
+    const sourceFiles = await getTrackedSourceFiles(root.path, logger, [
+      '.js',
+      '.mjs',
+      '.cjs',
+      '.jsx',
+      '.tsx'
+    ])
 
     expect(sourceFiles.map((file) => file.fileName)).toStrictEqual([
       path.join(root.path, 'test.cjs'),
@@ -31,7 +37,13 @@ describe('getTrackedSourceFiles', () => {
   it('correctly includes root', async () => {
     const root = new Fixture('projects/all-extensions/test.js')
 
-    const sourceFiles = await getTrackedSourceFiles(root.path, logger)
+    const sourceFiles = await getTrackedSourceFiles(root.path, logger, [
+      '.js',
+      '.mjs',
+      '.cjs',
+      '.jsx',
+      '.tsx'
+    ])
 
     expect(sourceFiles.map((file) => file.fileName)).toStrictEqual([root.path])
   })
@@ -39,7 +51,13 @@ describe('getTrackedSourceFiles', () => {
   it('correctly returns empty tracked files array when no files match the desired extensions', async () => {
     const root = new Fixture('projects/all-extensions/nested/deeply-nested/irrelevant-nested-files')
 
-    const sourceFiles = await getTrackedSourceFiles(root.path, logger)
+    const sourceFiles = await getTrackedSourceFiles(root.path, logger, [
+      '.js',
+      '.mjs',
+      '.cjs',
+      '.jsx',
+      '.tsx'
+    ])
 
     expect(sourceFiles.map((file) => file.fileName)).toStrictEqual([])
   })
@@ -47,7 +65,13 @@ describe('getTrackedSourceFiles', () => {
   it('correctly returns empty tracked files array when directory does not exist', async () => {
     const root = new Fixture('scopes/jsx/not-a-real-directory')
 
-    const sourceFiles = await getTrackedSourceFiles(root.path, logger)
+    const sourceFiles = await getTrackedSourceFiles(root.path, logger, [
+      '.js',
+      '.mjs',
+      '.cjs',
+      '.jsx',
+      '.tsx'
+    ])
 
     expect(sourceFiles.map((file) => file.fileName)).toStrictEqual([])
   })
