+++ src/test/scopes/js/node-handlers/import-node-handler.test.ts
@@ -7,9 +7,10 @@
 import * as ts from 'typescript'
 import { describe, expect, it } from 'vitest'
 
+import { getTrackedSourceFiles } from '../../../../main/core/get-tracked-source-files.js'
 import { ImportNodeHandler } from '../../../../main/scopes/js/node-handlers/import-node-handler.js'
 import { JsxElementAccumulator } from '../../../../main/scopes/jsx/jsx-element-accumulator.js'
-import { getTrackedSourceFiles } from '../../../../main/scopes/jsx/utils/get-tracked-source-files.js'
+import { JsxScope } from '../../../../main/scopes/jsx/jsx-scope.js'
 import { findNodesByType } from '../../../__utils/find-nodes-by-type.js'
 import { Fixture } from '../../../__utils/fixture.js'
 import { initLogger } from '../../../__utils/init-logger.js'
@@ -17,7 +18,9 @@ import { initLogger } from '../../../__utils/init-logger.js'
 describe('class: ImportNodeHandler', async () => {
   const logger = initLogger()
   const fixture = new Fixture('jsx-samples/all-import-types.tsx')
-  const sourceFile = (await getTrackedSourceFiles(fixture.path, logger))[0] as ts.SourceFile
+  const sourceFile = (
+    await getTrackedSourceFiles(fixture.path, logger, JsxScope.fileExtensions)
+  )[0] as ts.SourceFile
   const accumulator = new JsxElementAccumulator()
   const handler = new ImportNodeHandler(sourceFile, logger)
   it('correctly returns the JsxImports for a complex fixture', () => {
