+++ src/test/scopes/js/node-handlers/value-handlers/identifier-handler.test.ts
@@ -7,9 +7,10 @@
 import * as ts from 'typescript'
 import { describe, expect, it } from 'vitest'
 
+import { getTrackedSourceFiles } from '../../../../../main/core/get-tracked-source-files.js'
 import { ComplexValue } from '../../../../../main/scopes/js/complex-value.js'
 import { IdentifierHandler } from '../../../../../main/scopes/js/node-handlers/value-handlers/identifier-handler.js'
-import { getTrackedSourceFiles } from '../../../../../main/scopes/jsx/utils/get-tracked-source-files.js'
+import { JsxScope } from '../../../../../main/scopes/jsx/jsx-scope.js'
 import { findNodesByType } from '../../../../__utils/find-nodes-by-type.js'
 import { Fixture } from '../../../../__utils/fixture.js'
 import { initLogger } from '../../../../__utils/init-logger.js'
@@ -19,7 +20,9 @@ describe('identifierHandler', () => {
 
   it('correctly returns undefined for undefined attribute value', async () => {
     const fixture = new Fixture('jsx-samples/all-attr-types.tsx')
-    const sourceFile = (await getTrackedSourceFiles(fixture.path, logger))[0] as ts.SourceFile
+    const sourceFile = (
+      await getTrackedSourceFiles(fixture.path, logger, JsxScope.fileExtensions)
+    )[0] as ts.SourceFile
 
     const handler = new IdentifierHandler(sourceFile, logger)
 
@@ -34,7 +37,9 @@ describe('identifierHandler', () => {
 
   it('correctly returns complex attribute', async () => {
     const fixture = new Fixture('jsx-samples/all-attr-types.tsx')
-    const sourceFile = (await getTrackedSourceFiles(fixture.path, logger))[0] as ts.SourceFile
+    const sourceFile = (
+      await getTrackedSourceFiles(fixture.path, logger, JsxScope.fileExtensions)
+    )[0] as ts.SourceFile
 
     const handler = new IdentifierHandler(sourceFile, logger)
