+++ src/test/scopes/js/node-handlers/value-handlers/jsx-expression-handler.test.ts
@@ -7,9 +7,10 @@
 import * as ts from 'typescript'
 import { describe, expect, it } from 'vitest'
 
+import { getTrackedSourceFiles } from '../../../../../main/core/get-tracked-source-files.js'
 import { NoAttributeExpressionFoundError } from '../../../../../main/exceptions/no-attribute-expression-found-error.js'
 import { JsxExpressionHandler } from '../../../../../main/scopes/js/node-handlers/value-handlers/jsx-expression-handler.js'
-import { getTrackedSourceFiles } from '../../../../../main/scopes/jsx/utils/get-tracked-source-files.js'
+import { JsxScope } from '../../../../../main/scopes/jsx/jsx-scope.js'
 import { findNodesByType } from '../../../../__utils/find-nodes-by-type.js'
 import { Fixture } from '../../../../__utils/fixture.js'
 import { initLogger } from '../../../../__utils/init-logger.js'
@@ -18,7 +19,9 @@ describe('class: JsxExpressionHandler', () => {
   const logger = initLogger()
   it('correctly returns node text', async () => {
     const fixture = new Fixture('jsx-samples/all-attr-types.tsx')
-    const sourceFile = (await getTrackedSourceFiles(fixture.path, logger))[0] as ts.SourceFile
+    const sourceFile = (
+      await getTrackedSourceFiles(fixture.path, logger, JsxScope.fileExtensions)
+    )[0] as ts.SourceFile
     const handler = new JsxExpressionHandler(sourceFile, logger)
 
     const nodes = findNodesByType<ts.JsxExpression>(sourceFile, ts.SyntaxKind.JsxExpression)
@@ -29,7 +32,9 @@ describe('class: JsxExpressionHandler', () => {
 
   it('throws NoAttributeExpressionFoundError if attribute does not have an expression', async () => {
     const fixture = new Fixture('jsx-samples/all-jsx-element-types.tsx')
-    const sourceFile = (await getTrackedSourceFiles(fixture.path, logger))[0] as ts.SourceFile
+    const sourceFile = (
+      await getTrackedSourceFiles(fixture.path, logger, JsxScope.fileExtensions)
+    )[0] as ts.SourceFile
     const handler = new JsxExpressionHandler(sourceFile, logger)
 
     const node = findNodesByType(
