+++ src/test/scopes/jsx/node-handlers/elements/jsx-element-node-handler.test.ts
@@ -7,9 +7,10 @@
 import * as ts from 'typescript'
 import { describe, expect, it } from 'vitest'
 
+import { getTrackedSourceFiles } from '../../../../../main/core/get-tracked-source-files.js'
 import { JsxElementAccumulator } from '../../../../../main/scopes/jsx/jsx-element-accumulator.js'
+import { JsxScope } from '../../../../../main/scopes/jsx/jsx-scope.js'
 import { JsxElementNodeHandler } from '../../../../../main/scopes/jsx/node-handlers/elements/jsx-element-node-handler.js'
-import { getTrackedSourceFiles } from '../../../../../main/scopes/jsx/utils/get-tracked-source-files.js'
 import { findNodesByType } from '../../../../__utils/find-nodes-by-type.js'
 import { Fixture } from '../../../../__utils/fixture.js'
 import { initLogger } from '../../../../__utils/init-logger.js'
@@ -17,7 +18,9 @@ import { initLogger } from '../../../../__utils/init-logger.js'
 describe('class: JsxElementNodeHandler', async () => {
   const logger = initLogger()
   const fixture = new Fixture('jsx-samples/all-jsx-element-types.tsx')
-  const sourceFile = (await getTrackedSourceFiles(fixture.path, logger))[0] as ts.SourceFile
+  const sourceFile = (
+    await getTrackedSourceFiles(fixture.path, logger, JsxScope.fileExtensions)
+  )[0] as ts.SourceFile
   const accumulator = new JsxElementAccumulator()
   const handler = new JsxElementNodeHandler(sourceFile, logger)
   it('correctly returns the JsxElements for a complex fixture', async () => {
